Refactor Et variable to float

+++ cuda/src/include/utils.h
@@ -28,8 +28,8 @@ void read_data(char filename[], float *x, double *p, unsigned int *evolution, in
 
 void save_data(char filename[], float *x, double *p, unsigned int evolution, int N_PART);
 
-double energy_sum(double *p, int N_PART, unsigned int evolution, double M);
+float energy_sum(double *p, int N_PART, unsigned int evolution, double M);
 
-int make_hist(int *h_h, int *h_g, int *h_hg, int *d_h, int *d_g, int *d_hg, double *DxE, double *DpE, const char *filename, int BINS, double Et);
+int make_hist(int *h_h, int *h_g, int *h_hg, int *d_h, int *d_g, int *d_hg, double *DxE, double *DpE, const char *filename, int BINS, float Et);
 
 #endif
