+++ cuda/src/utils/utils.cu
@@ -64,7 +64,7 @@ void read_data(char filename[], float *h_x, double *h_p, unsigned int *evolution
     fclose(readFile);
 }
 
-double energy_sum(double *p, int N_PART, unsigned int evolution, double M) {
+float energy_sum(double *p, int N_PART, unsigned int evolution, double M) {
     double *d_p, *d_partialSum;
     double *partialSum;
     int threadsPerBlock = 256; // Define the number of threads per block
@@ -100,7 +100,7 @@ double energy_sum(double *p, int N_PART, unsigned int evolution, double M) {
 
     // Output the result
     std::cout << "N° de pasos " << evolution << "\tEnergía total = " << sumEnergy / (2 * M) << std::endl;
-    return sumEnergy / (2 * M);
+    return (float)(sumEnergy / (2 * M));
 }
 
 void save_data(char filename[], float *x, double *p, unsigned int evolution, int N_PART) {
@@ -163,74 +163,6 @@ void save_data(char filename[], float *x, double *p, unsigned int evolution, int
     saveFile.write(reinterpret_cast<const char *>(p), sizeof(p[0]) * N_PART);
 }
 
-// int make_hist(int *h, int *g, int *hg, double *DxE, double *DpE, const char *filename, int BINS, double Et) {
-//     double chi2x = 0.0, chi2xr = 0.0, chi2p = 0.0, chiIp = 0.0, chiPp = 0.0, chiIx = 0.0, chiPx = 0.0;
-
-//     if (strcmp(filename, "X0000000.dat") == 0) {
-//         #pragma omp parallel for reduction(+ : chi2x) schedule(static)
-//         for (int i = BINS ; i < 2 * BINS; i++) {
-//             chi2x += pow(h[i] - 2 * DxE[i], 2) / (2 * DxE[i]);
-//         }
-//         chi2x /= BINS;
-//     } else {
-//         #pragma omp parallel for reduction(+ : chi2x) schedule(static)
-//         for (int i = 4; i < 2 * BINS; i++) {
-//             chi2x += pow(h[i] - DxE[i], 2) / DxE[i];
-//         }
-//         chi2x /= (2.0 * BINS - 4);
-//         chi2xr = chi2x; // chi2xr = chi2x reducido
-//     }
-    
-//     #pragma omp parallel for reduction(+ : chi2p) schedule(static)
-//     for (int i = 0; i < 2 * (BINS - BORDES); i++) {
-//         chi2p += pow(g[i + BORDES] - DpE[i + BORDES], 2) / DpE[i + BORDES];
-//     }
-    
-//     #pragma omp parallel for reduction(+ : chiIp, chiPp) schedule(static)
-//     for (int i = 0; i < (BINS - BORDES); i++) {
-//         chiIp += pow(g[i + BORDES] - g[2 * BINS - 1 - BORDES - i], 2) / DpE[i + BORDES];
-//         chiPp += pow(g[i + BORDES] + g[2 * BINS - 1 - BORDES - i] - 2.0 * DpE[i + BORDES], 2) / DpE[i + BORDES];
-//     }
-
-//     #pragma omp parallel for reduction(+ : chiIx, chiPx) schedule(static)
-//     for (int i = 4; i <= BINS + 1; i++) {
-//         chiIx += pow(h[i] - h[2 * BINS + 3 - i], 2) / DxE[i];
-//         chiPx += pow(h[i] + h[2 * BINS + 3 - i] - 2.0 * DxE[i], 2) / DxE[i];
-//     }
-
-//     chiIx = chiIx / (2.0 * (BINS - 2));
-//     chiPx = chiPx / (2.0 * (BINS - 2));
-//     chi2p = chi2p / (2.0 * (BINS - BORDES));
-//     chiIp = chiIp / (2.0 * (BINS - BORDES));
-//     chiPp = chiPp / (2.0 * (BINS - BORDES));
-
-//     FILE *hist = fopen(filename, "w");
-//     fprintf(hist,
-//             "#   x    poblacion       p      poblacion    chi2x =%9.6f  chi2xr "
-//             "=%9.6f  chiIx =%9.6f  chiPx =%9.6f  chi2p =%9.6f  chiIp =%9.6f  "
-//             "chiPp =%9.6f  Et=%12.9E\n",
-//             chi2x, chi2xr, chiIx, chiPx, chi2p, chiIp, chiPp, Et);
-//     cout << "#   x    poblacion       p      poblacion    chi2x =" << chi2x << "  chi2xr =" << chi2xr << "  chiIx ="
-//          << chiIx << "  chiPx =" << chiPx << "  chi2p =" << chi2p << "  chiIp =" << chiIp << "  chiPp =" << chiPp
-//          << "  Et=" << Et << endl;
-//     fprintf(hist, "%8.5f %6d %24.12E %6d\n", -0.5015, h[0], -2.997e-23, g[0]);
-//     fprintf(hist, "%8.5f %6d %24.12E %6d\n", -0.5005, h[1], -2.997e-23, g[0]);
-//     for (int i = 0; i < BINS << 1; i++) {
-//         fprintf(hist, "%8.5f %6d %24.12E %6d\n", (0.5 * i / BINS - 0.4995), h[i + 2], (3.0e-23 * i / BINS - 2.997e-23),
-//                 g[i]);
-//     }
-//     fprintf(hist, "%8.5f %6d %24.12E %6d\n", 0.5005, h[2 * BINS + 2], 2.997e-23, g[2 * BINS - 1]);
-//     fprintf(hist, "%8.5f %6d %24.12E %6d\n", 0.5015, h[2 * BINS + 3], 2.997e-23, g[2 * BINS - 1]);
-
-//     fclose(hist);
-
-//     memset(h, 0, (2 * BINS + 4) * sizeof(int));
-//     memset(g, 0, (2 * BINS) * sizeof(int));
-//     memset(hg, 0, (2 * BINS + 4) * (2 * BINS) * sizeof(int));
-
-//     return 0; // avisa que se cumplió la condición sobre los chi2
-// }
-
 __global__ void chi2x_kernel(int *h, double *DxE, double *chi2x, int BINS, bool isX0000000) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     double local_chi2x = 0.0;
@@ -280,7 +212,7 @@ __global__ void chiIx_Px_kernel(int *h, double *DxE, double *chiIx, double *chiP
 }
 
 
-int make_hist(int *h_h, int *h_g, int *h_hg, int *d_h, int *d_g, int *d_hg, double *DxE, double *DpE, const char *filename, int BINS, double Et) {
+int make_hist(int *h_h, int *h_g, int *h_hg, int *d_h, int *d_g, int *d_hg, double *DxE, double *DpE, const char *filename, int BINS, float Et) {
     double *d_chi2x, *d_chi2p, *d_chiIp, *d_chiPp, *d_chiIx, *d_chiPx;
 
     // Allocate memory for reduction variables on GPU
